requirements.txt
fastapi
uvicorn
spacy
sentence-transformers
torch
python-docx
requests
python-dotenv